test: split job addressing tests for hinting and indexing
